backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-multipart
Pillow
rapidfuzz
requests
python-dotenv
easyocr
inference-sdk
